feat(onboarding): SIF immediate indexing, step 3 semantic insights, PlatformAnalytics extraction
Part A — Immediate SIF indexing after website analysis:
- Add _immediate_sif_index() async helper in onboarding_task_scheduler.py
- Wire asyncio.create_task into complete_step() after schedule_step2_tasks
- Best-effort with scheduler SIFIndexingTask fallback

Part B — SIF semantic intelligence in Step 3:
- Add semantic_insights, content_analysis, strategic_recommendations to CompetitorDiscoveryResponse
- Call enhance_step3_with_semantic_intelligence alongside traditional discovery
- Graceful degradation: SIF fields are null on error

Frontend SIF UI:
- Extract and store SIF fields in CompetitorAnalysisStep state+cache
- New AI Semantic Insights section (Content Pillars, Content Gaps, Recommended Actions)
- Update AnalysisProgressDisplay scheduling message for indexing

PlatformAnalytics extraction:
- Extract MetricsCard, AIInsightsPanel, BingToolbar, CannibalizationAlertsPanel,
  ContentBriefDialog, SummaryCharts as self-contained components
- Fix duplicate keyframe animation names, null guards
- Fix Fade import in WebsiteStep
